Extract `TestCase` object to its own module
Reclaims the `saulify.sitespec` module for structures exclusively
related to the spec files themselves.
